Most people exhibit what political scientists call "the conservatism of the peasantry."  Don't lose what you've got.  Don't change.  Don't take a chance, because you might end up starving to death.  Play it safe.  Buy just as much as you need.  Don't waste time.
When  we think about risk, human beings and corporations realize in their
heads that risks are necesary to grow, to survive.  But when it comes down
to keeping good people when the crunch comes, or investing money in
something untried, only the brave reach deep into their pockets and play
the game as it must be played.

- David Lammers, "Yakitori", Electronic Engineering Times, January 18, 1988